Make workspace repositories code base maintainable (prerelease)
