+++ CLAUDE.md
@@ -46,8 +46,8 @@ CI pushes to Docker Hub on commits to `main`, on tag pushes, or on any branch co
 
 GrayMoon is a **two-process** system — never combine them:
 
-- **GrayMoon.App** (`src/GrayMoon.App`) — ASP.NET Core 8 + Blazor Server, runs in Docker. Exposes the web UI and two SignalR hubs (`/hub/agent` for App↔Agent commands, `/hubs/workspace-sync` for browser broadcast). Never touches the local filesystem or runs git directly. SQLite database via EF Core at `/app/db` (volume-mounted). Uses `EnsureCreated()` — no incremental migrations. **New DB columns must be nullable or have a default value.**
-- **GrayMoon.Agent** (`src/GrayMoon.Agent`) — .NET console app / tool that runs on the developer's host machine. Connects to the App via SignalR, executes all git and filesystem operations, and exposes a local HTTP listener on `127.0.0.1:9191` for git hook callbacks. Job queue: bounded channel with up to 16 concurrent workers.
+- **GrayMoon.App** (`src/GrayMoon.App`) — ASP.NET Core **.NET 8** + Blazor Server, runs in Docker. Exposes the web UI and two SignalR hubs (`/hub/agent` for App↔Agent commands, `/hubs/workspace-sync` for browser broadcast). Never touches the local filesystem or runs git directly. SQLite database via EF Core at `/app/db` (volume-mounted). Uses `EnsureCreated()` — no incremental migrations. **New DB columns must be nullable or have a default value.**
+- **GrayMoon.Agent** (`src/GrayMoon.Agent`) — .NET **10** console app / dotnet tool (`graymoon-agent` command) that runs on the developer's host machine. Packaged as a single binary; supports Windows Service (`Microsoft.Extensions.Hosting.WindowsServices`) and Linux Systemd (`Microsoft.Extensions.Hosting.Systemd`) hosting. CLI entry via `AgentCli.Build()` with subcommands: `run` (default), `install`, `uninstall`, `start`, `stop`. Connects to the App via SignalR, executes all git and filesystem operations, and exposes a local HTTP listener on `127.0.0.1:9191` for git hook callbacks. Job queue: bounded channel with up to 16 concurrent workers. Three hosted services manage the Agent lifecycle: `SignalRConnectionHostedService` (maintains hub connection), `HookListenerHostedService` (HTTP listener for git hook POSTs), and `JobBackgroundService` (dequeues and executes jobs).
 - **GrayMoon.Abstractions** (`src/GrayMoon.Abstractions`) — Shared interfaces and DTOs used by both App and Agent. `Agent/AgentHubMethods.cs` contains all SignalR hub method name constants (`RequestCommand`, `ResponseCommand`, `SyncCommand`, `CommandOutput`, `ReportSemVer`, `ReportQueueStatus`).
 - **GrayMoon.Common** (`src/GrayMoon.Common`) — Shared utilities including `CommandLineService` (process execution wrapper) and `Search/FilterSearchExpression` (boolean expression parser used by every UI grid).
