+++ src/GrayMoon.App/Models/DependencyLine.cs
@@ -1,0 +1,4 @@
+namespace GrayMoon.App.Models;
+
+/// <summary>Describes a single package dependency of a workspace repository (package ID and its current version).</summary>
+public sealed record DependencyLine(string PackageId, string Version);
